here I am adding the old stuff to the new project
Here I am adding the old stuff to the new project so we can experiment with it and use it if needed for the new stuff coming. I tried to use or new organization system but some stuff was way to much work to organize. but is almost the same!